[web] Move sendTextMessage to ChatInputStateContainer
