+++ web/chat/failed-send.react.js
@@ -6,15 +6,11 @@ import {
 } from 'lib/selectors/chat-selectors';
 import {
   messageTypes,
-  type SendMessageResult,
-  type SendMessagePayload,
   type RawComposableMessageInfo,
-  type RawTextMessageInfo,
   assertComposableMessageType,
 } from 'lib/types/message-types';
 import { type ThreadInfo, threadInfoPropType } from 'lib/types/thread-types';
 import type { AppState } from '../redux-setup';
-import type { DispatchActionPromise } from 'lib/utils/action-utils';
 import {
   chatInputStatePropType,
   type ChatInputState,
@@ -26,13 +22,10 @@ import PropTypes from 'prop-types';
 
 import { messageID } from 'lib/shared/message-utils';
 import { connect } from 'lib/utils/redux-utils';
-import {
-  sendTextMessageActionTypes,
-  sendTextMessage,
-} from 'lib/actions/message-actions';
 
 import css from './chat-message-list.css';
 import multimediaMessageSendFailed from './multimedia-message-send-failed';
+import textMessageSendFailed from './text-message-send-failed';
 
 type Props = {|
   item: ChatMessageInfoItem,
@@ -40,14 +33,6 @@ type Props = {|
   chatInputState: ChatInputState,
   // Redux state
   rawMessageInfo: RawComposableMessageInfo,
-  // Redux dispatch functions
-  dispatchActionPromise: DispatchActionPromise,
-  // async functions that hit server APIs
-  sendTextMessage: (
-    threadID: string,
-    localID: string,
-    text: string,
-  ) => Promise<SendMessageResult>,
 |};
 class FailedSend extends React.PureComponent<Props> {
   static propTypes = {
@@ -55,29 +40,49 @@ class FailedSend extends React.PureComponent<Props> {
     threadInfo: threadInfoPropType.isRequired,
     chatInputState: chatInputStatePropType.isRequired,
     rawMessageInfo: PropTypes.object.isRequired,
-    dispatchActionPromise: PropTypes.func.isRequired,
-    sendTextMessage: PropTypes.func.isRequired,
   };
   retryingText = false;
   retryingMedia = false;
 
   componentDidUpdate(prevProps: Props) {
-    const isFailed = multimediaMessageSendFailed(
-      this.props.item,
-      this.props.chatInputState,
-    );
-    const wasFailed = multimediaMessageSendFailed(
-      prevProps.item,
-      prevProps.chatInputState,
-    );
-    const isDone =
-      this.props.item.messageInfo.id !== null &&
-      this.props.item.messageInfo.id !== undefined;
-    const wasDone =
-      prevProps.item.messageInfo.id !== null &&
-      prevProps.item.messageInfo.id !== undefined;
-    if ((isFailed && !wasFailed) || (isDone && !wasDone)) {
-      this.retryingMedia = false;
+    if (
+      (this.props.rawMessageInfo.type === messageTypes.IMAGES ||
+        this.props.rawMessageInfo.type === messageTypes.MULTIMEDIA) &&
+      (prevProps.rawMessageInfo.type === messageTypes.IMAGES ||
+        prevProps.rawMessageInfo.type === messageTypes.MULTIMEDIA)
+    ) {
+      const isFailed = multimediaMessageSendFailed(
+        this.props.item,
+        this.props.chatInputState,
+      );
+      const wasFailed = multimediaMessageSendFailed(
+        prevProps.item,
+        prevProps.chatInputState,
+      );
+      const isDone =
+        this.props.item.messageInfo.id !== null &&
+        this.props.item.messageInfo.id !== undefined;
+      const wasDone =
+        prevProps.item.messageInfo.id !== null &&
+        prevProps.item.messageInfo.id !== undefined;
+      if ((isFailed && !wasFailed) || (isDone && !wasDone)) {
+        this.retryingMedia = false;
+      }
+    } else if (
+      this.props.rawMessageInfo.type === messageTypes.TEXT &&
+      prevProps.rawMessageInfo.type === messageTypes.TEXT
+    ) {
+      const isFailed = textMessageSendFailed(this.props.item);
+      const wasFailed = textMessageSendFailed(prevProps.item);
+      const isDone =
+        this.props.item.messageInfo.id !== null &&
+        this.props.item.messageInfo.id !== undefined;
+      const wasDone =
+        prevProps.item.messageInfo.id !== null &&
+        prevProps.item.messageInfo.id !== undefined;
+      if ((isFailed && !wasFailed) || (isDone && !wasDone)) {
+        this.retryingText = false;
+      }
     }
   }
 
@@ -100,13 +105,11 @@ class FailedSend extends React.PureComponent<Props> {
       if (this.retryingText) {
         return;
       }
-      const newRawMessageInfo = { ...rawMessageInfo, time: Date.now() };
-      this.props.dispatchActionPromise(
-        sendTextMessageActionTypes,
-        this.sendTextMessageAction(newRawMessageInfo),
-        undefined,
-        newRawMessageInfo,
-      );
+      this.retryingText = true;
+      this.props.chatInputState.sendTextMessage({
+        ...rawMessageInfo,
+        time: Date.now(),
+      });
     } else if (
       rawMessageInfo.type === messageTypes.IMAGES ||
       rawMessageInfo.type === messageTypes.MULTIMEDIA
@@ -120,36 +123,6 @@ class FailedSend extends React.PureComponent<Props> {
       this.props.chatInputState.retryMultimediaMessage(localID);
     }
   };
-
-  async sendTextMessageAction(
-    messageInfo: RawTextMessageInfo,
-  ): Promise<SendMessagePayload> {
-    this.retryingText = true;
-    try {
-      const { localID } = messageInfo;
-      invariant(
-        localID !== null && localID !== undefined,
-        'localID should be set',
-      );
-      const result = await this.props.sendTextMessage(
-        messageInfo.threadID,
-        localID,
-        messageInfo.text,
-      );
-      return {
-        localID,
-        serverID: result.id,
-        threadID: messageInfo.threadID,
-        time: result.time,
-      };
-    } catch (e) {
-      e.localID = messageInfo.localID;
-      e.threadID = messageInfo.threadID;
-      throw e;
-    } finally {
-      this.retryingText = false;
-    }
-  }
 }
 
 export default connect(
@@ -167,5 +140,4 @@ export default connect(
     );
     return { rawMessageInfo };
   },
-  { sendTextMessage },
 )(FailedSend);
